video6 cambio de byte a int

+++ CursoJava/src/app/Variables.java
@@ -6,7 +6,7 @@ package app;
 public class Variables {
 
     public static void main(String[] args) {
-        byte edad;
+        int edad;
         edad = 35;
         System.out.println(edad);
     }
